LayoutTests/canvas/philip/tests/2d.imageData.get.unaffected-expected.txt

< [index] >
2d.imageData.get.unaffected
getImageData() is not affected by context state
References: 2d.imageData.unaffected
Actual output:
Expected output:

Passed
